title: 60fps via interpolation at ORIGINAL logic rate — revert the R_UPDATE_RATE=1 override (#149)
User feedback: rate=1 made the demo (moon/sky) run too fast; the intended
architecture is original 20fps logic + InterpolationFPS=60 presentation.

- revert the title R_UPDATE_RATE override (logic back to R_UPDATE_RATE=3,
  cs cadence back to its pre-change rate); keep InterpolationFPS forced to 60
- key the sun/moon draw with FrameInterpolation_RecordOpenChild: at root level
  its matrices only pair positionally between logic frames, and op drift broke
  the pairing ~every 6th present (measured moon holds + 1.5px catch-ups while
  the world interpolated smoothly)
- Zelda3D_TitleCsTimeOfDay now takes a fractional cs frame (integer cue latch,
  fractional +6/frame drift) so dayTime-derived sun/moon/dome/light state
  advances every engine frame instead of stepping per cs tick

Measured: logic 19.9fps, presents 59.7fps (interpolation live); moon centroid
motion smooth except a ~1px 1-in-6 residual (dayTime quantization) now also
addressed by the fractional dayTime.

+++ Shipwright/soh/src/zelda3d/cutscene/zelda3d_cutscene.h
@@ -47,7 +47,9 @@ int Zelda3D_TitleCsRiderCue(int frame, int* cueIndex,
 
 // Time-of-day from op-0x8c cues (latched; title = 4:01 AM). Returns 0
 // before the first cue frame.
-int Zelda3D_TitleCsTimeOfDay(int frame, uint16_t* outDayTime);
+// Fractional frame (f + Zelda3D_TitleCsSubframe()): cue latch is integer, the +6/cs-frame drift
+// advances fractionally so dayTime-derived state moves every engine frame (#149).
+int Zelda3D_TitleCsTimeOfDay(float frame, uint16_t* outDayTime);
 
 // spot99's scene light settings (raw ZSI cmd-0x0F 28-byte entries).
 int Zelda3D_TitleCsLightSlotsRaw(const uint8_t** outSlots, int* outCount);
